requirements.txt
Django
django_debug_toolbar
djangorestframework
djangorestframework-simplejwt
dj-database-url
psycopg2
psycopg2-binary
python-decouple
django-cors-headers
whitenoise
django-filter
pytest
pytest-django
factory-boy
pytest-mock
Faker
django-cloudinary-storage
cloudinary
Pillow
drf-spectacular
python-dotenv
gunicorn
redis


